requirements.txt
# GUI Framework
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0

# EXIF Processing
Pillow>=10.0.0
exifread>=3.0.0

# OCR
paddleocr>=2.7.0
paddlepaddle>=2.5.0

# Database
pyodbc>=5.0.0
openpyxl>=3.1.0
pandas>=2.1.0

# Utilities
python-dateutil>=2.8.0
python-docx>=1.2.0

# Logging
colorlog>=6.7.0
